backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
qrcode[pil]==7.4.2
anthropic==0.39.0
ccxt==4.4.13
celery[redis]==5.4.0
redis[hiredis]==5.1.1
redbeat==2.2.0
pandas==2.2.2
ta==0.11.0
httpx==0.27.2
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.2
websockets==14.1
cryptography==43.0.1
